pre call refactor
