+++ template/src/components/precall/PreCallSettings.tsx
@@ -47,6 +47,7 @@ const styles = StyleSheet.create({
     borderRadius: 4,
     padding: 0,
     width: '100%',
+    height: '95%',
   },
   containerStyle: {},
 });
