+++ template/src/components/precall/VideoPreview.native.tsx
@@ -12,12 +12,10 @@
 
 import {useContent, useRtc} from 'customization-api';
 import React from 'react';
-import {View, StyleSheet} from 'react-native';
 import {MaxVideoView} from '../../../agora-rn-uikit';
-import PreCallLocalMute from './LocalMute';
-import {ImageIcon as UiKitImageIcon} from '../../../agora-rn-uikit';
+import VideoFallback from './VideoFallback';
 
-const VideoPreview: React.FC = () => {
+const VideoPreview = () => {
   const rtc = useRtc();
   rtc?.RtcEngineUnsafe?.startPreview();
 
@@ -29,52 +27,19 @@ const VideoPreview: React.FC = () => {
   }
 
   return (
-    <View style={styles.container}>
-      <View style={{flex: 1}}>
-        <MaxVideoView
-          user={defaultContent[maxUid]}
-          key={maxUid}
-          fallback={Fallback}
-          containerStyle={{
-            width: '100%',
-            height: '100%',
-            borderBottomLeftRadius: 8,
-            borderBottomRightRadius: 8,
-          }}
-        />
-      </View>
-      <PreCallLocalMute isMobileView={true} />
-    </View>
+    <MaxVideoView
+      user={defaultContent[maxUid]}
+      key={maxUid}
+      fallback={VideoFallback}
+      containerStyle={{
+        width: '100%',
+        height: '100%',
+        borderTopLeftRadius: 8,
+        borderTopRightRadius: 8,
+      }}
+      isPrecallScreen={true}
+    />
   );
 };
+
 export default VideoPreview;
-
-const Fallback = () => {
-  return (
-    <View style={styles.fallbackRootContainer}>
-      <View style={styles.avatar}>
-        {/*TODO fix ttf file <ImageIcon name="profile" customSize={{width: 100, height: 100}} /> */}
-        <UiKitImageIcon name={'profile'} />
-      </View>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    position: 'relative',
-    justifyContent: 'space-between',
-  },
-  avatar: {
-    height: 100,
-    width: 100,
-  },
-  fallbackRootContainer: {
-    flex: 1,
-    backgroundColor: $config.VIDEO_AUDIO_TILE_COLOR,
-    justifyContent: 'center',
-    alignItems: 'center',
-    position: 'relative',
-  },
-});
